requirements.txt
flask==3.1.0
openai==1.66.0  # Updated to a more recent version
httpx==0.28.1   # Specify compatible httpx version
anthropic==0.49.0
google-generativeai==0.8.4
python-dotenv==1.0.1
